+++ src/heap.c
@@ -1064,12 +1064,11 @@ gboolean
 heap_string_new(Heap *heap, const char *str, PElement *out)
 {
 	PElement list = *out;
-	const int n = strlen(str);
-	int i;
 
 	heap_list_init(&list);
 
-	for (i = 0; i < n; i++) {
+	const int n = strlen(str);
+	for (int i = 0; i < n; i++) {
 		PElement t;
 
 		if (!heap_list_add(heap, &list, &t))
@@ -1088,7 +1087,8 @@ heap_managedstring_new(Heap *heap, const char *str, PElement *out)
 {
 	Managedstring *managedstring;
 
-	if (strcmp(str, "") == 0) {
+	if (!str ||
+		g_str_equal(str, "")) {
 		PEPUTP(out, ELEMENT_ELIST, NULL);
 	}
 	else {
@@ -2435,7 +2435,9 @@ lisp_pelement(VipsBuf *buf, PElement *base,
 		break;
 
 	case ELEMENT_MANAGED:
-		vips_buf_appendf(buf, "<Managed* %p>", PEGETVAL(base));
+		vips_buf_appendf(buf, "<Managed* %p, ", PEGETVAL(base));
+		iobject_info(IOBJECT(PEGETVAL(base)), buf);
+		vips_buf_appendf(buf, ">");
 		break;
 
 	case ELEMENT_COMB: